Move test to tests/rt_handle.rs
