+++ tokio/tests/rt_handle.rs
@@ -3,6 +3,8 @@
 #![cfg(feature = "full")]
 
 use tokio::runtime::Runtime;
+use tokio::sync::{mpsc, Barrier};
+use std::sync::Arc;
 
 #[test]
 #[cfg_attr(panic = "abort", ignore)]
@@ -63,6 +65,40 @@ fn interleave_then_enter() {
     // Can still enter
     let rt3 = rt();
     let _enter = rt3.enter();
+}
+
+// If the cycle causes a leak, then miri will catch it.
+#[test]
+fn drop_tasks_with_reference_cycle() {
+    rt().block_on(async {
+        let (tx, mut rx) = mpsc::channel(1);
+
+        let barrier = Arc::new(Barrier::new(3));
+        let barrier_a = barrier.clone();
+        let barrier_b = barrier.clone();
+
+        let a = tokio::spawn(async move {
+            let b = rx.recv().await.unwrap();
+
+            // Poll the JoinHandle once. This registers the waker.
+            // The other task cannot have finished at this point due to the barrier below.
+            futures::future::select(b, std::future::ready(())).await;
+
+            barrier_a.wait().await;
+        });
+
+        let b = tokio::spawn(async move {
+            // Poll the JoinHandle once. This registers the waker.
+            // The other task cannot have finished at this point due to the barrier below.
+            futures::future::select(a, std::future::ready(())).await;
+
+            barrier_b.wait().await;
+        });
+
+        tx.send(b).await.unwrap();
+
+        barrier.wait().await;
+    });
 }
 
 #[cfg(tokio_unstable)]
